requirements.txt
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Flask==1.1.1
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
newsapi-python==0.2.6
requests==2.23.0
urllib3==1.25.8
Werkzeug==1.0.0
